Refactored tsa-cli
