+++ tsa-cli/src/main/kotlin/org/ton/options/FiftOptions.kt
@@ -1,0 +1,14 @@
+package org.ton.options
+
+import com.github.ajalt.clikt.parameters.groups.OptionGroup
+import com.github.ajalt.clikt.parameters.options.help
+import com.github.ajalt.clikt.parameters.options.option
+import com.github.ajalt.clikt.parameters.options.required
+import com.github.ajalt.clikt.parameters.types.path
+
+class FiftOptions : OptionGroup("Fift options") {
+    val fiftStdlibPath by option("--fift-std")
+        .path(mustExist = true, canBeDir = true, canBeFile = false)
+        .required()
+        .help("The path to the Fift standard library (dir containing Asm.fif, Fift.fif)")
+}
